+++ internal/lnd/client.go
@@ -133,13 +133,15 @@ func (c *Client) SubscribeEvents() (<-chan events.Event, error) {
 	}
 
 	// Start subscription handlers
+	// NOTE: Keep handlers in alphabetical order to prevent merge conflicts when adding new handlers
 	handlers := []func(){
-		c.handleForwards,
-		c.handlePeerEvents,
 		c.handleChannelEvents,
-		c.handleInvoiceEvents,
 		c.handleFailedHtlcEvents,
+		c.handleForwards,
+		c.handleInvoiceEvents,
 		c.handleKeysendEvents,
+		c.handlePaymentEvents,
+		c.handlePeerEvents,
 	}
 	c.wg.Add(len(handlers))
 	for _, h := range handlers {